injection_monitor/requirements.txt
requests
feedparser
selenium
python-dotenv
pywin32
schedule
beautifulsoup4